Pairing screen edit

+++ src/remote/remote.h
@@ -152,7 +152,7 @@ String BOARDMENUS[mainBoardMenus] = {"Board 1", "Board 2", "Board 3", "Board 4",
 
 enum menu_boards { BOARDS_SELECT, BOARDS_DELETE };
 
-long boards[20] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
+uint32_t boards[20] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
 char* conversionChart[20] = {"0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19"};
 
 int selectedBoardSlot = 0;
